+++ features/Context/NavigationContext.php
@@ -81,18 +81,9 @@ class NavigationContext extends RawMinkContext implements PageObjectAwareInterfa
      */
     public function resetCurrentPage()
     {
-        $this->getMainContext()->executeScript('sessionStorage.clear();');
+        $script = 'sessionStorage.clear(); typeof $ !== "undefined" && $(window).off("beforeunload");';
+        $this->getMainContext()->executeScript($script);
         $this->currentPage = null;
-    }
-
-    /**
-     * @AfterScenario
-     */
-    public function disableNavigationConfirmation()
-    {
-        if (strpos($this->currentPage, 'edit') || strpos($this->currentPage, 'creation')) {
-            $this->getMainContext()->executeScript('typeof $ !== "undefined" && $(window).off("beforeunload");');
-        }
     }
 
     /**
@@ -102,11 +93,8 @@ class NavigationContext extends RawMinkContext implements PageObjectAwareInterfa
      */
     public function iAmLoggedInAs($username)
     {
-        $password = $username;
-        $this->getFixturesContext()->getOrCreateUser($username, $password);
-
         $this->username = $username;
-        $this->password = $password;
+        $this->password = $username;
     }
 
     /**